Folder and Files renaming and reorganization refactor
