+++ la-gloria-store/src/pages/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductTable from '../HomeComponents/ProductTable/ProductTable';
+import ProductTable from '../components/HomePageComponents/ProductTable';
 import '../App.css';
